+++ src/test/kotlin/step1/Person.kt
@@ -1,3 +1,3 @@
-package study.step1
+package step1
 
 data class Person(val name: String, val age: Int, var nickname: String? = name)
