+++ src/test/kotlin/step1/PersonKoTest.kt
@@ -1,4 +1,4 @@
-package study.step1
+package step1
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
