+++ src/test/kotlin/step1/StringTest.kt
@@ -1,4 +1,4 @@
-package study.step1
+package step1
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
