+++ src/test/kotlin/step2/StringCalculatorTest.kt
@@ -1,7 +1,6 @@
-package study
+package step2
 
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatExceptionOfType
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
